+++ run_bot_webhook.py
@@ -1,12 +1,12 @@
 import uvicorn
-
-from fastapi import FastAPI, Request
-from telegram import Update
+from fastapi import FastAPI
 
 from src.main import start_bot, create_bot
+from src.api.routers import webhook_router
 
 
 app = FastAPI()
+app.include_router(webhook_router)
 
 
 @app.on_event("startup")
@@ -21,17 +21,6 @@ async def on_shutdown():
     """Действия после остановки сервера с ботом."""
     bot_app = create_bot()
     await bot_app.shutdown()
-
-
-@app.post('/telegram_handler')
-async def get_telegram_bot_updates(request: Request) -> dict:
-    """Получение обновлений telegram в режиме работы бота webhook."""
-    bot_app = request.app.state.bot_app
-    request_json_data = await request.json()
-    await bot_app.update_queue.put(
-        Update.de_json(data=request_json_data, bot=bot_app.bot)
-    )
-    return request_json_data
 
 
 if __name__ == '__main__':
