+++ src/main.py
@@ -3,13 +3,16 @@ from urllib.parse import urljoin
 from fastapi import FastAPI
 from telegram.ext import ApplicationBuilder, CommandHandler, Application
 
-from src.api.routers import router
+from src.api.routers import router, webhook_router, WEBHOOK_ENDPOINT
 from src.bot.handlers import start
 from src.core.settings import settings
 
 app = FastAPI()
 
 app.include_router(router)
+
+if settings.BOT_WEBHOOK_MODE:
+    app.include_router(webhook_router)
 
 
 def create_bot() -> Application:
@@ -30,7 +33,7 @@ async def bot_init_webhook_mode(bot_app: Application) -> None:
     """Инициализировать бота в режиме webhook"""
     bot_app.updater = None
     await bot_app.bot.set_webhook(
-        url=urljoin(settings.APPLICATION_URL, 'telegram_handler'),
+        url=urljoin(settings.APPLICATION_URL, WEBHOOK_ENDPOINT),
         secret_token=settings.BOT_TOKEN.replace(':', '')  # colon is not allowed here # noqa
     )
     await bot_app.initialize()
